+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 
 ## Unreleased
 
+## [1.1.0](https://github.com/dbmdz/mirador-textoverlay/releases/tag/1.1.0) - 2026-08-14
+
 ### Added
 
 - Support for IIIFv3 annotations and OCR references in `rendering` and `seeAlso`